Extract inline class to `PostgresSimulationResultsHandle`
